Rename task_test.rs
